requirements.txt
beautifulsoup4==4.13.5
celery==5.5.3
celery-types==0.23.0
lxml==6.0.2
psycopg2==2.9.10
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.5
selenium==4.35.0
SQLAlchemy==2.0.43
